Remove the WinUI file
