+++ source/SkiaSharp.Views.WinUI/SkiaSharp.Views.WinUI/SkiaSharp.Views.WinUI.csproj
@@ -18,8 +18,8 @@
   <ItemGroup>
     <Compile Include="..\..\SkiaSharp.Views\SkiaSharp.Views.Shared\**\*.cs" Link="%(RecursiveDir)%(Filename)%(Extension)" />
     <Compile Include="..\..\SkiaSharp.Views\SkiaSharp.Views.UWP\**\*.cs" Link="%(RecursiveDir)%(Filename)%(Extension)" Exclude="..\..\SkiaSharp.Views\SkiaSharp.Views.UWP\bin\**;..\..\SkiaSharp.Views\SkiaSharp.Views.UWP\obj\**" />
+    <Compile Remove="..\..\SkiaSharp.Views\SkiaSharp.Views.UWP\Interop\IBufferByteAccess.cs" />
   </ItemGroup>
-  
   <ItemGroup Condition="'$(BuildingInsideVisualStudio)' != 'true'">
     <ProjectReference Include="..\..\..\binding\SkiaSharp.NativeAssets.WinUI\SkiaSharp.NativeAssets.WinUI.csproj" />
   </ItemGroup>
